cliente_api_transporte/requirements.txt
flask
pyodbc
requests
python-dotenv